+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 
   - Major/Minor/Patch: This is an example changelog item, usually can be the commit message.
   - Append new items to make git merging easier.
+  - Patch: Ensure Currency component can update its value given new `value` prop
+  - Minor: Add `onChange` prop to CustomFieldInputText component
+  - Minor: Add `onChange` prop to CustomFieldInputNumber component
+  - Minor: Add `onChange` prop to CustomFieldInputCurrency component
 </details>
 
 ## v0.53.3